requirements.txt
certifi==2020.12.5
numpy==1.21.0
opencv-python==4.5.3.56
pandas==1.3.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
wincertstore==0.2
pyyaml==5.4.1